Fix issues with running sources live when other endpoints don't exist yet
