+++ frontend/portal/src/api/sources.ts
@@ -1,23 +1,61 @@
 import { httpJson } from "@portal/api/http";
-import type { Source, SourcesResponse } from "@portal/mocks/sources";
 
 /**
  * Sources service layer: the backend contract.
- *
- * Like the policies surface (and unlike the portal's mock-only `/v1/...` pages),
- * this calls the REAL Stirling base `/api/v1/sources` (SourceController), so it is
- * plug-and-play: drop MSW and these exact calls hit the live backend.
  */
 
-export type {
-  Source,
-  SourceDetailRow,
-  SourceKpi,
-  SourcePolicyRef,
-  SourceStatus,
-  SourceView,
-  SourcesResponse,
-} from "@portal/mocks/sources";
+/** Overview row status: referenced and enabled, enabled-but-orphaned, or disabled. */
+export type SourceStatus = "active" | "unused" | "disabled";
+
+/** A policy that references a source. */
+export interface SourcePolicyRef {
+  id: string;
+  name: string;
+}
+
+/** One key/value line summarising a source's config for display. */
+export interface SourceDetailRow {
+  label: string;
+  value: string;
+}
+
+/** One row in the Sources overview table. Mirrors the backend `SourceView`. */
+export interface SourceView {
+  id: string;
+  name: string;
+  type: string;
+  status: SourceStatus;
+  referenceCount: number;
+  referencingPolicies: SourcePolicyRef[];
+  config: SourceDetailRow[];
+  /** Per-source document volume — not tracked yet (always null for now). */
+  docsTotal: number | null;
+}
+
+export interface SourceKpi {
+  value: number;
+  description: string;
+}
+
+export interface SourcesResponse {
+  kpis: SourceKpi[];
+  sources: SourceView[];
+}
+
+/**
+ * The wire record for a single source: the create/update body (`id` omitted on
+ * create) and what the backend returns from POST/GET. Mirrors the backend
+ * `Source` record; `owner`/`teamId` are stamped server-side.
+ */
+export interface Source {
+  id?: string;
+  name: string;
+  type: string;
+  options: Record<string, unknown>;
+  enabled: boolean;
+  owner?: string | null;
+  teamId?: number | null;
+}
 
 /** GET /api/v1/sources: KPI strip + one row per source for the admin. */
 export async function fetchSources(): Promise<SourcesResponse> {
